requirements.txt
aiohttp
discord.py
bs4